refactored: person-card folder name
