+++ src/components/person-details/person-card/CardAvatar.tsx
