+++ src/components/person-details/person-card/CardBackground.tsx
